Simple Linked List Implementation in C
This C program provides a basic implementation of a singly linked list. It includes functions for common linked list operations such as:

Inserting at the beginning: Adds a new node at the start of the list.
Inserting at the end: Appends a new node at the end of the list.
Deleting a node by value: Removes the first node that matches a specified value.
Printing the list: Displays the current nodes in the linked list in sequence.